+++ components/layout.tsx
@@ -1,6 +1,6 @@
-import Footer from '@/components/footer'
-import Header from '@/components/header'
-import Main from '@/components/main'
+import Footer from '@/components/layout-components/footer'
+import Header from '@/components/layout-components/header'
+import Main from '@/components/layout-components/main'
 import { useLocale } from '@/lib/hooks'
 import theme from '@/lib/theme.preval'
 import React from 'react'
