+++ components/onezero-skip-to-main-content.tsx
@@ -1,7 +1,7 @@
 export default function OneZeroSkipToMainContent({
   text,
   dir,
-  className,
+  className = '',
   style,
 }: {
   text: string
