+++ pages/index.tsx
@@ -23,8 +23,8 @@ export default function Home({ data }: HomePageType) {
       <AboutTheClinic data={data} />
       <WhatWeDo data={data} />
       <ContactUs data={data} />
-      Here is the structure of the home page data:
-      <pre className="bg-dark text-light">{JSON.stringify(data, null, 2)}</pre>
+      {/* Here is the structure of the home page data:
+      <pre className="bg-dark text-light">{JSON.stringify(data, null, 2)}</pre> */}
     </>
   )
 }
